% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERSPCdiagnosisdeathAndTreatedRightways.R
\docType{package}
\name{ERSPCdiagnosisdeathAndTreatedRightways}
\alias{ERSPCdiagnosisdeathAndTreatedRightways}
\alias{ERSPCdiagnosisdeathAndTreatedRightways-package}
\title{ERSPCdiagnosisdeathAndTreatedRightways}
\description{
ERSPCdiagnosisdeathAndTreatedRightways
}
